:clown_face: Barebones blog site with sample hello world post
Creating a fairly minimal static site in eleventy. With one index page and one hello-world post! Writeup readme (for github) on how to replicate this setup, with a pretty directory structure.

The post list is based off https://v0-5-0.11ty.io/docs/collections/#a-blog-example, with an <a href> tag added to navigate to the hello-world post.

Permalink syntax follows that of https://v0-5-0.11ty.io/docs/permalinks/#use-filters!